Ajustes generales migración a la nube

+++ src/main/java/co/com/segurosalfa/siniestros/repo/ISnrAbogadoRepo.java
@@ -2,7 +2,7 @@ package co.com.segurosalfa.siniestros.repo;
 
 import co.com.segurosalfa.siniestros.entity.SnrAbogado;
 
-public interface ISnrAbogadoRepo extends IGenericRepo<SnrAbogado, Long> {
+public interface ISnrAbogadoRepo extends IGenericRepo<SnrAbogado, Integer> {
 
 
 }
