+++ src/main/java/co/com/segurosalfa/siniestros/service/IAbogadoService.java
@@ -2,6 +2,6 @@ package co.com.segurosalfa.siniestros.service;
 
 import co.com.segurosalfa.siniestros.entity.SnrAbogado;
 
-public interface IAbogadoService extends ICRUD<SnrAbogado, Long>{
+public interface IAbogadoService extends ICRUD<SnrAbogado, Integer>{
 
 }
